added hit detection for hero and enemy, and moved notes and maps to external files
