+++ maps.js
@@ -1,0 +1,27 @@
+let maps = [
+    {
+        id: 1,
+        walls: [
+            {
+                type: 'vertical',
+                length: 4,
+                startX: 3,
+                startY: 2
+            },
+            {
+                type: 'horizontal',
+                length: 4,
+                startX: 4,
+                startY: 1
+            },
+            {
+                type: 'vertical',
+                length: 3,
+                startX: 7,
+                startY: 5
+            }
+        ]
+    }
+];
+
+export { maps }
